Issue452-ScoreGraphChanges
In this commit requested changes were made to the Score out of 100 graph in the results page. The changes included: colors of score bars to match the colors of the progress bars and the back ground of score graphs to be a single shade of gray instead of white and blue stripes. For more information refer to #Issue452.